ref(timeseries): Move timeseries code to base
- This code was identical in every dataset except for the
  comparison_delta stuff, which isn't run when the param is None anyways
  so feels pretty safe
- The only dataset that never implemented timeseries was uptime_results,
  so keeping that the same just to keep this as no-op as possible
- Also minor cleanup in the endpoint code
