+++ src/sentry/snuba/profile_functions.py
@@ -1,17 +1,13 @@
 import logging
-from datetime import timedelta
 
 import sentry_sdk
 from sentry_protos.snuba.v1.request_common_pb2 import PageToken
 
 from sentry.search.eap.profile_functions.definitions import PROFILE_FUNCTIONS_DEFINITIONS
 from sentry.search.eap.resolver import SearchResolver
-from sentry.search.eap.sampling import events_meta_from_rpc_request_meta
 from sentry.search.eap.types import AdditionalQueries, EAPResponse, SearchResolverConfig
-from sentry.search.events.types import SAMPLING_MODES, EventsMeta, SnubaParams
+from sentry.search.events.types import SAMPLING_MODES, SnubaParams
 from sentry.snuba import rpc_dataset_common
-from sentry.snuba.discover import zerofill
-from sentry.utils.snuba import SnubaTSResult
 
 logger = logging.getLogger("sentry.snuba.profile_functions")
 
@@ -58,65 +54,4 @@ class ProfileFunctions(rpc_dataset_common.RPCBase):
                 additional_queries=additional_queries,
             ),
             debug=debug,
-        )
-
-    @classmethod
-    @sentry_sdk.trace
-    def run_timeseries_query(
-        cls,
-        *,
-        params: SnubaParams,
-        query_string: str,
-        y_axes: list[str],
-        referrer: str,
-        config: SearchResolverConfig,
-        sampling_mode: SAMPLING_MODES | None,
-        comparison_delta: timedelta | None = None,
-        additional_queries: AdditionalQueries | None = None,
-    ) -> SnubaTSResult:
-        cls.validate_granularity(params)
-        search_resolver = cls.get_resolver(params, config)
-        rpc_request, aggregates, groupbys = cls.get_timeseries_query(
-            search_resolver=search_resolver,
-            params=params,
-            query_string=query_string,
-            y_axes=y_axes,
-            groupby=[],
-            referrer=referrer,
-            sampling_mode=sampling_mode,
-            additional_queries=additional_queries,
-        )
-
-        """Run the query"""
-        rpc_response = cls._run_timeseries_rpc(params.debug, rpc_request)
-
-        """Process the results"""
-        result = rpc_dataset_common.ProcessedTimeseries()
-        final_meta: EventsMeta = events_meta_from_rpc_request_meta(rpc_response.meta)
-        for resolved_field in aggregates + groupbys:
-            final_meta["fields"][resolved_field.public_alias] = resolved_field.search_type
-
-        for timeseries in rpc_response.result_timeseries:
-            processed = cls.process_timeseries_list([timeseries])
-            if len(result.timeseries) == 0:
-                result = processed
-            else:
-                for attr in ["timeseries", "confidence", "sample_count", "sampling_rate"]:
-                    for existing, new in zip(getattr(result, attr), getattr(processed, attr)):
-                        existing.update(new)
-        if len(result.timeseries) == 0:
-            # The rpc only zerofills for us when there are results, if there aren't any we have to do it ourselves
-            result.timeseries = zerofill(
-                [],
-                params.start_date,
-                params.end_date,
-                params.timeseries_granularity_secs,
-                ["time"],
-            )
-
-        return SnubaTSResult(
-            {"data": result.timeseries, "processed_timeseries": result, "meta": final_meta},
-            params.start,
-            params.end,
-            params.granularity_secs,
         )
